split up properties handling for service configuration
